fix issue #109 - refactor key fill/scan to not crash after pruning
After a prune, key fill/scan can crash because there is an assumption
that data is always resident in the cache for a small key.

This code handles the case where the small key data was evicted during
a prune, and reads the data back in just for that key on demand.

Furthermore, during a `sync` call, it will pull in the key data
that corresponds to the page containing just the small key, allowing
that page to be re-composed and replaced in memory. caching is used
on the decryption of the small key page so the expensive bit
of the call should only happen once per small key bank.

tested with pddb-ci and manually in hosted mode.

A portion of code was re-written to make it clear that only the
length of allocated data is copied to/from disk and memory, instead
of the reserved space as the previous code would indicate. The reserved
space is guaranteed to be zero because the RAM-based page for composing
the small key block is zero-ized prior to operations; left comments
reminding me of this in the future so I don't accidentally optimize
something out and cause data leakage between small keys because I'm
allowing the reserved areas to contain stale data.