add bulk response chaining
IN descriptors now chain up to 512 bytes long

OUT descriptors still need to be worked up

However, there is an issue where the system is injecting the
ZLP/CSW immediately after the bulk response is queued up, and this
is getting sent before the bulk data is issued. Need to figure out
this race condition.