change suspend timeout behavior
When suspend fails to happen within the designated time,
instead of forcing it, return a failure code so we can retry it
again later.

In testing, this seems to be a very reasonable approach, because
sometimes servers are just too busy to do a suspend, and just
waiting a couple seconds until eg some file commits are done
makes everything work much better.