codec: place codec object into heap
The codec object contains a lot of data including a `FrameRing` buffer
as well as some other services. With the latest version of Rust, this
has grown large and no longer fits onto the stack. As a result, it
crashes with a stack overflow immediately upon boot.

Move the Codec onto the heap. This frees up stack and allows it to run
under Rust 1.69.0.

Signed-off-by: Sean Cross <sean@xobs.io>