rework list filtering to use an unsorted vec instead of btreemap
also eliminates references, copying and locking of items in the
sorted btreemap.

turns out that ... sorting is much cheaper than copying or maintaining
references, because copies to heap items require allocs which are
under the hoood several sorts to find the place to stick the alloc
memory, and shared references require locks.