Allow logging from spawn
This allows the spawn (i.e. new process) to use log. Firstly,
xous-api-log was changed to allow no_std when the target os is
xous (this is done using a feature). Then, in spawn itself, an
AtomicU32 is moved into memory as a replacement for
XOUS_LOGGER_CONNECTION, which is not in .text and is thus not in the
stub file.