Working + somewhat improved UI!
`app-loader' can now successfully download apps from the app server
and run them! I have also improved the UI by adding a progressbar
when an app is loading so that you know when it is done and also
replacing some panics with notifications that way the whole machine
doesn't shut down if the server has some problems.

The remaining UI work is to make sure that the menu pops back up when
you for example set the server.

I also need to add locales, a README, etc.